moved the analytics page
