--- src/app/VolunteerNavView/Analytics/page.module.css
@@ -1,44 +1,0 @@
-.container {
-    display: flex;
-    height: 100vh;
-  }
-  
-  .sidebar {
-    width: 250px;
-  }
-  
-  .mainContent {
-    flex: 1;
-    padding: 2rem;
-  }
-  
-  .header {
-    display: flex;
-    flex-direction: column; 
-    align-items: flex-start; 
-    gap: 0.5rem;
-    margin-bottom: 1rem;
-  }
-  
-  .backButton {
-    background: none;
-    border: none;
-    font-size: 1rem;
-    color: #64748B; 
-    cursor: pointer;
-    display: flex;
-    align-items: center;
-    gap: 4px; 
-  }
-
-  .arrow {
-    font-size: 1rem;
-  }
-  
-  .pageTitle {
-    margin: 0;
-    font-size: 1.5rem;
-    font-weight: 600; 
-    color: #000000;
-  }
-
